rename modal to CreateColumn
